dual_radio.c for z1 updated, should update dual_radio.c for zoul

+++ core/net/rpl/rpl-icmp6.c
@@ -355,7 +355,7 @@ dis_output(uip_ipaddr_t *addr)
    *     +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
    */
 	char *log_buf = (char*) malloc(sizeof(char)*100);
-	sprintf(log_buf,"DIS_OUTPUT, Energy: %d\n",get_residual_energy()); 
+	sprintf(log_buf,"DIS_OUTPUT, Energy: %d\n",(int)get_residual_energy()); 
 	dis_count ++;
 	LOG_MESSAGE(log_buf); 
 	free(log_buf);
@@ -689,7 +689,7 @@ dio_output(rpl_instance_t *instance, uip_ipaddr_t *uc_addr)
 #endif /* !RPL_LEAF_ONLY */
 
 	char *log_buf = (char*) malloc(sizeof(char)*100);
-	sprintf(log_buf,"DIO_OUTPUT, Energy: %d\n",get_residual_energy()); 
+	sprintf(log_buf,"DIO_OUTPUT, Energy: %d\n",(int) get_residual_energy()); 
 	dio_count ++;
 	LOG_MESSAGE(log_buf); 
 	free(log_buf);
@@ -1099,7 +1099,7 @@ dio_ack_output(rpl_instance_t *instance, uip_ipaddr_t *uc_addr)
 	rpl_parent_t *p = dag->preferred_parent;
 
 	char *log_buf = (char*) malloc(sizeof(char)*100);
-	sprintf(log_buf,"DIO_ACK_OUTPUT, Energy: %d\n",get_residual_energy());
+	sprintf(log_buf,"DIO_ACK_OUTPUT, Energy: %d\n",(int) get_residual_energy());
 	dio_ack_count ++;
 	LOG_MESSAGE(log_buf);
 	free(log_buf);
@@ -1630,7 +1630,7 @@ dao_output(rpl_parent_t *parent, uint8_t lifetime)
   uip_ipaddr_t prefix;
 
 	char *log_buf = (char*) malloc(sizeof(char)*100);
-	sprintf(log_buf,"DAO_OUTPUT, Energy: %d\n",get_residual_energy()); 
+	sprintf(log_buf,"DAO_OUTPUT, Energy: %d\n",(int) get_residual_energy()); 
 	dao_count ++;
 	LOG_MESSAGE(log_buf); 
 	free(log_buf);
